endpoint: move getEndpointList from daemon to EndpointManager
The method `Daemon.getEndpointList` is currently used by the
Endpoint API and the Debug API.

In preparation for extracting the Endpoint API, this commit moves
the method from the `Daemon` to the `EndpointManager`.
