+++ daemon/cmd/endpoint.go
@@ -10,7 +10,6 @@ import (
 	"maps"
 	"net"
 	"net/http"
-	"runtime"
 	"sync"
 
 	"github.com/go-openapi/runtime/middleware"
@@ -52,7 +51,7 @@ func getEndpointHandler(d *Daemon, params GetEndpointParams) middleware.Responde
 	}
 	defer r.Done()
 
-	resEPs := d.getEndpointList(params)
+	resEPs := d.endpointManager.GetEndpointList(params)
 
 	if params.Labels != nil && len(resEPs) == 0 {
 		r.Error(errEndpointNotFound, GetEndpointNotFoundCode)
@@ -60,67 +59,6 @@ func getEndpointHandler(d *Daemon, params GetEndpointParams) middleware.Responde
 	}
 
 	return NewGetEndpointOK().WithPayload(resEPs)
-}
-
-func (d *Daemon) getEndpointList(params GetEndpointParams) []*models.Endpoint {
-	maxGoroutines := runtime.NumCPU()
-	var (
-		epWorkersWg, epsAppendWg sync.WaitGroup
-		convertedLabels          labels.Labels
-		resEPs                   []*models.Endpoint
-	)
-
-	if params.Labels != nil {
-		// Convert params.Labels to model that we can compare with the endpoint's labels.
-		convertedLabels = labels.NewLabelsFromModel(params.Labels)
-	}
-
-	eps := d.endpointManager.GetEndpoints()
-	if len(eps) < maxGoroutines {
-		maxGoroutines = len(eps)
-	}
-	epsCh := make(chan *endpoint.Endpoint, maxGoroutines)
-	epModelsCh := make(chan *models.Endpoint, maxGoroutines)
-
-	epWorkersWg.Add(maxGoroutines)
-	for range maxGoroutines {
-		// Run goroutines to process each endpoint and the corresponding model.
-		// The obtained endpoint model is sent to the endpoint models channel from
-		// where it will be aggregated later.
-		go func(wg *sync.WaitGroup, epModelsChan chan<- *models.Endpoint, epsChan <-chan *endpoint.Endpoint) {
-			for ep := range epsChan {
-				if ep.HasLabels(convertedLabels) {
-					epModelsChan <- ep.GetModel()
-				}
-			}
-			wg.Done()
-		}(&epWorkersWg, epModelsCh, epsCh)
-	}
-
-	// Send the endpoints to be aggregated a models to the endpoint channel.
-	go func(epsChan chan<- *endpoint.Endpoint, eps []*endpoint.Endpoint) {
-		for _, ep := range eps {
-			epsChan <- ep
-		}
-		close(epsChan)
-	}(epsCh, eps)
-
-	epsAppendWg.Add(1)
-	// This needs to be done over channels since we might not receive all
-	// the existing endpoints since not all endpoints contain the list of
-	// labels that we will use to filter in `ep.HasLabels(convertedLabels)`
-	go func(epsAppended *sync.WaitGroup) {
-		for ep := range epModelsCh {
-			resEPs = append(resEPs, ep)
-		}
-		epsAppended.Done()
-	}(&epsAppendWg)
-
-	epWorkersWg.Wait()
-	close(epModelsCh)
-	epsAppendWg.Wait()
-
-	return resEPs
 }
 
 func deleteEndpointHandler(d *Daemon, params DeleteEndpointParams) middleware.Responder {
